tests: merge test/data-hook into test/hook-override
