--- test/data-hook/.testCiJob.json
@@ -1,1 +1,0 @@
-{ "name": "Unit Tests E2E" }
