--- test/data-hook/package.json
@@ -1,18 +1,0 @@
-{
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "preview": "vite build && vite preview"
-  },
-  "dependencies": {
-    "@types/react": "^18.2.15",
-    "@types/react-dom": "^18.2.7",
-    "@vitejs/plugin-react": "^3.0.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "typescript": "^5.1.3",
-    "vite": "^4.0.3",
-    "vike": "0.4.147"
-  },
-  "type": "module"
-}
