--- test/data-hook/pages/+config.h.ts
@@ -1,7 +1,0 @@
-export default {
-  // prettier-ignore
-  passToClient: [
-    'globalOnBeforeRenderWasCalled',
-    'globalOnBeforeRenderWasCalledInEnv'
-  ]
-}
