--- test/data-hook/pages/page-2/+onBeforeRender.tsx
@@ -1,1 +1,0 @@
-export default null
