--- test/data-hook/pages/page-3/+Page.tsx
@@ -1,1 +1,0 @@
-../index/+Page.tsx
