--- test/data-hook/pages/page-4/+Page.tsx
@@ -1,1 +1,0 @@
-../index/+Page.tsx
