--- test/data-hook/pages/page-4/+meta.ts
@@ -1,8 +1,0 @@
-export default {
-  data: {
-    env: { server: false, client: true }
-  },
-  onBeforeRender: {
-    env: { server: true, client: true }
-  }
-}
