--- test/data-hook/renderer/+config.h.ts
@@ -1,4 +1,0 @@
-export default {
-  clientRouting: true,
-  hydrationCanBeAborted: true
-}
