--- test/data-hook/renderer/+onRenderHtml.tsx
@@ -1,26 +1,0 @@
-export default onRenderHtml
-
-import ReactDOMServer from 'react-dom/server'
-import React from 'react'
-import { escapeInject, dangerouslySkipEscape } from 'vike/server'
-import { PageLayout } from './PageLayout'
-import type { PageContextServer } from 'vike/types'
-
-function onRenderHtml(pageContext: PageContextServer) {
-  const { Page } = pageContext
-  const pageHtml = ReactDOMServer.renderToString(
-    <PageLayout pageContext={pageContext}>
-      <Page />
-    </PageLayout>
-  )
-
-  return escapeInject`<!DOCTYPE html>
-    <html>
-      <head>
-        <meta charset="UTF-8" />
-      </head>
-      <body>
-        <div id="page-view">${dangerouslySkipEscape(pageHtml)}</div>
-      </body>
-    </html>`
-}
