--- test/data-hook/renderer/PageLayout.css
@@ -1,7 +1,0 @@
-body {
-  margin: 0;
-  font-family: sans-serif;
-}
-* {
-  box-sizing: border-box;
-}
