--- test/data-hook/renderer/usePageContext.tsx
@@ -1,16 +1,0 @@
-export { PageContextProvider }
-export { usePageContext }
-
-import React, { useContext } from 'react'
-import type { PageContext } from 'vike/types'
-
-const Context = React.createContext<PageContext>(undefined as any)
-
-function PageContextProvider({ pageContext, children }: { pageContext: PageContext; children: React.ReactNode }) {
-  return <Context.Provider value={pageContext}>{children}</Context.Provider>
-}
-
-function usePageContext() {
-  const pageContext = useContext(Context)
-  return pageContext
-}
