--- test/data-hook/test-dev.test.ts
@@ -1,2 +1,0 @@
-import { test } from './test'
-test('npm run dev')
