--- test/data-hook/vite.config.ts
@@ -1,7 +1,0 @@
-import react from '@vitejs/plugin-react'
-import vike from 'vike/plugin'
-import type { UserConfig } from 'vite'
-
-export default {
-  plugins: [react(), vike()]
-} satisfies UserConfig
