+++ test/hook-override/pages/+config.h.ts
@@ -1,8 +1,9 @@
 export default {
   // prettier-ignore
   passToClient: [
-    'onBeforeRender1WasCalled',
-    'onBeforeRender2WasCalled',
-    'onBeforeRenderEnv'
+    'globalOnBeforeRenderWasCalled',
+    'globalOnBeforeRenderWasCalledInEnv',
+    'perPageOnBeforeRenderWasCalled',
+    'perPageOnBeforeRenderWasCalledInEnv'
   ]
 }
