+++ test/hook-override/pages/+onBeforeRender.tsx
@@ -1,8 +1,13 @@
-export default () => {
+// https://vike.dev/onBeforeRender
+export { onBeforeRender }
+
+import type { OnBeforeRenderAsync } from 'vike/types'
+
+const onBeforeRender: OnBeforeRenderAsync = async (pageContext): ReturnType<OnBeforeRenderAsync> => {
   return {
     pageContext: {
-      onBeforeRender1WasCalled: true,
-      onBeforeRenderEnv: typeof window === 'undefined' ? 'server' : 'client'
+      globalOnBeforeRenderWasCalled: true,
+      globalOnBeforeRenderWasCalledInEnv: typeof window === 'undefined' ? 'server' : 'client'
     }
   }
 }
