+++ test/hook-override/pages/page-2/+config.ts
