--- test/hook-override/pages/page-3/+config.ts
@@ -1,3 +1,0 @@
-export default {
-  onBeforeRender: null
-}
