+++ test/hook-override/pages/page-3/README.md
@@ -1,3 +1,3 @@
 This page:
-* overrides the `data()` hook via `+data.tsx`
+* overrides the global `data()` hook via `+data.tsx`
 * doesn't override the global `onBeforeRender()` hook
