+++ test/hook-override/pages/page-4/README.md
@@ -1,5 +1,5 @@
 This page:
-* overrides the `data()` hook via `+data.tsx`
+* overrides the global `data()` hook via `+data.tsx`
 * changes the `data()` hook's `env` to client only via `+meta.ts`
-* doesn't override the global `onBeforeRender()` hook
+* overrides the global `onBeforeRender()` hook via `+onBeforeRender.tsx`
 * changes the `onBeforeRender()` hook's `env` to client and server via `+meta.ts`
