+++ test/hook-override/renderer/PageLayout.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import './PageLayout.css'
 import { PageContextProvider } from './usePageContext'
-import type { PageContext } from './types'
+import type { PageContext } from 'vike/types'
 
 export { PageLayout }
