+++ test/hook-override/tsconfig.json
@@ -1,7 +1,13 @@
 {
   "compilerOptions": {
     "strict": true,
+    "module": "ES2020",
+    "moduleResolution": "Node",
+    "target": "ES2020",
+    "lib": ["DOM", "DOM.Iterable", "ESNext"],
+    "types": ["vite/client"],
     "jsx": "react",
+    "skipLibCheck": true,
     "esModuleInterop": true
   }
 }
